Fixed UI element themes across multiple screens and controllers
- Updated UI components in `calendar_screen.dart`, `dashboard_screen.dart`, `habits_screen.dart`, `home_screen.dart`, and `profile_screen.dart`.
- Adjusted related logic in `calendar_controller.dart`, `dashboard_controller.dart`, `habits_controller.dart`, and `profile_controller.dart`.
- Added new models: `achievements.dart`, `activity_log.dart`, `habit.dart`, `habit_category.dart`, `habit_model.dart`, and `habit_record.dart`.
- Minor updates to `main.dart` and `api_service.dart`.
