+++ habitflow3/lib/models/achievements.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class Achievement {
+  final String id;
+  final String title;
+  final String description;
+  final IconData icon;
+  final DateTime earnedAt;
+
+  Achievement({
+    required this.id,
+    required this.title,
+    required this.description,
+    required this.icon,
+    required this.earnedAt,
+  });
+}
