+++ habitflow3/lib/models/habit_category.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+
+class HabitCategory {
+  final String name;
+  final Color color;
+  final IconData icon;
+
+  HabitCategory({
+    required this.name,
+    required this.color,
+    required this.icon,
+  });
+}
